+++ test/kat/Pipfile
@@ -5,6 +5,8 @@ verify_ssl = true
 
 [dev-packages]
 autopep8 = "*"
+flake8 = "*"
+flake8-mypy = "*"
 
 [packages]
 pytest = "*"
